Add Unix::Telnet and Windows::Telnet
Net::Telnet used to be on this list, but was removed after v0.0.1 was released. However, it doesn't implement terminal emulation, which is needed for a proper TELNET implementation, since it's platform dependent. Two more modules need to be created before it can be completely removed from the most wanted modules list.